shared: fix keyring handling in ask-password-api
Commit 74dd6b515fa968c5710b396a7664cac335e25ca8 (core: run each system
service with a fresh session keyring) broke keyring handling in ask-password.
The key was added to user keyring, but permissions denied to access it. So
instead add the key to session keyring, set permission, link to user keyring
and unlink from session keyring.

For whatever reason requesting a key does not longer work. Search instead.