core/automount: revise the "fun fact" comments.
Fun fact 1 suggests that a "close()" is needed, but that
close() has long since been removed.  So the comment
in now meaningless and possibly confusing.

Fun fact 2 refers to a bug that has been fixed in
Linux prior to v4.12

Commit: 9fa4eb8e490a ("autofs: sanity check status reported with AUTOFS_DEV_IOCTL_FAIL")

so revise the comment so that no-one goes pointlessly
looking for the bug.