journal: cache last accessed FileDescriptor *
We spend a lot of time in hashmap_get() mapping fd numbers to
FileDescriptor * objects.  It's quite common for the same fd to
be accessed repeatedly, so caching the last accessed fd is a very
simple way to cut down the number of hashmap_get() calls.

It probably makes sense to refactor the mmap cache such that the
callers can simply supply the FileDescriptor * instead of the fd
number, eliminating the fds HashMap altogether.

Results from a simple test after spamming the journal up to 80M
across 20 small 4M journal files:

Before:
 $ time ./journalctl -b --no-pager > /dev/null

 real    0m11.033s
 user    0m11.023s
 sys     0m0.010s
 $ time ./journalctl -b --no-pager > /dev/null

 real    0m11.075s
 user    0m11.065s
 sys     0m0.010s
 $ time ./journalctl -b --no-pager > /dev/null

 real    0m11.233s
 user    0m11.225s
 sys     0m0.008s

After:
 $ time ./journalctl -b --no-pager > /dev/null

 real    0m10.381s
 user    0m10.368s
 sys     0m0.010s
 $ time ./journalctl -b --no-pager > /dev/null

 real    0m10.435s
 user    0m10.421s
 sys     0m0.011s
 $ time ./journalctl -b --no-pager > /dev/null

 real    0m10.395s
 user    0m10.384s
 sys     0m0.008s

There's definitely substantial room for improvement still, but this
change is trivial.