unit: fail stop job and inform the client when control command fails
When user stops a service but it fails to stop properly due to ExecStop
failure we pretend everything is fine and user is not informed about the
fact. This commit introduces change in behavior so that user gets
error message on the command line when control command fails.

Example:
$ cat /etc/systemd/system/foo.service
[Service]
ExecStart=/bin/sleep 3600
ExecStop=/bin/bash -c 'killall sleep; exit 1;'

Old:
$ systemctl start foo.service
$ systemctl stop foo.service
$ echo $?
0

New:
$ systemctl start foo.service
$ systemctl stop foo.service
Job for foo.service failed because the control process exited with error code.
See "systemctl  status foo.service" and "journalctl  -xe" for details.
$ echo $?
1

Note that change affects only stop jobs. Similar behavior is undesirable
for restart jobs because systemd would not attempt to start the service
in case when stop part of the restart job failed. IOW, there is no
change in behavior for (try-)restart case.