journald: start versioning the stdout stream protocol
The existing protocol is considered as version=0. A version can be sent
as part of the header, as a numeric (integer) line right after the
stream identifier. If it isn't a number, we fall back to version=0 and
treat the second line as the unit id string, keeping backwards
compatibility.

For versions > 0, ignore all unknown fields at the bottom (end) of the
header, until a special STDOUT_STREAM_HEADER_MARKER line is seen, so in
the future new lines (fields) can be added to the protocol in a
backwards compatible way (allowing forwards compatibility from now on).