core: restore unreachable states MOUNT_REMOUNTING_SIG{TERM,KILL}
They were made unreachable by commit f8d0ceb4 "when reloading a service
fails don't fail the entire service but just the reload job".

Favour restoring the states (with behaviour modified to fit).  Remount is
likely subject to similar hangs as unmount.  In particular it should be
nice to have the warning message in the case SIGKILL doesn't clean it up.

Also update the timeout warnings for start and stop, for accuracy and
consistency with their equivalents in service.c.