Path-activate systemd-binfmt
The systemd-binfmt service is enabled by default, but with conditions
checking that certain directories are not empty.

Because on most Linux-based operating systems those folders are empty by
default, when an admin adds some file in one of them (for example by
installing the qemu-user-static package on Fedora), they must then
either manually start the service, or reboot the machine.

By path-activating the service, the service is started as soon as these
files appear, making the whole process even easier than it already is.